fix(i18n): isolate localized guardrail validation
